Refs #10079; Split large model ReferentialCopy
